+++ ChangeLog.md
@@ -473,6 +473,7 @@ These artifacts include extensions for the types available in the latter JDKs, s
 ##### New features
 
 - [`KT-15068`](https://youtrack.jetbrains.com/issue/KT-15068) Implement intention which rename file according to the top-level class name
+- Implement quickfix which enables/disables coroutine support in module or project
 
 ## 1.0.6